+++ lumicks/pylake/detail/widefield.py
@@ -4,6 +4,7 @@ import json
 import cv2
 import tifffile
 import warnings
+from copy import copy
 
 
 class TiffFrame:
@@ -18,49 +19,21 @@ class TiffFrame:
 
     def __init__(self, page, align):
         self._src = page
+        self._description = ImageDescription(page)
         self._align = align
 
     def _align_image(self):
         """ reconstruct image using alignment matrices from Bluelake; return aligned image as a NumPy array"""
 
-        def correct_alignment_offset(alignment, x_offset, y_offset):
-            # translate the origin of the image so that it matches that of the original transform
-            translation = np.eye(3)
-            translation[0, -1] = -x_offset
-            translation[1, -1] = -y_offset
-            # apply the original transform to the translated image.
-            # it only needs to be resized from a 2x3 to a 3x3 matrix
-            original = np.vstack((alignment, [0, 0, 1]))
-            # translate the image back to the original origin.
-            # takes into account both the offset and the scaling performed by the first step
-            back_translation = np.eye(3)
-            back_translation[0, -1] = original[0, 0] * x_offset
-            back_translation[1, -1] = original[1, 1] * y_offset
-            # concatenate the transforms by multiplying their matrices and ignore the unused 3rd row
-            return np.dot(back_translation, np.dot(original, translation))[:2, :]
-
-        try:
-            description = json.loads(self._src.description)
-        except json.decoder.JSONDecodeError:
+        if not self._description:
             warnings.warn("File does not contain metadata. Only raw data is available")
             return self.raw_data
 
         try:
-            align_mats = [
-                np.array(description[f"Alignment {color} channel"]).reshape((2, 3))
-                for color in ("red", "blue")
-            ]
-            align_roi = np.array(description["Alignment region of interest (x, y, width, height)"])[
-                :2
-            ]
-            roi = np.array(description["Region of interest (x, y, width, height)"])[:2]
+            align_mats = [self._description.alignment_matrix(color) for color in ("red", "blue")]
         except KeyError:
             warnings.warn("File does not contain alignment matrices. Only raw data is available")
             return self.raw_data
-
-        x_offset, y_offset = align_roi - roi
-        if not (x_offset == 0 and y_offset == 0):
-            align_mats = [correct_alignment_offset(mat, x_offset, y_offset) for mat in align_mats]
 
         img = self.raw_data
         rows, cols, _ = img.shape
@@ -153,3 +126,71 @@ class TiffStack:
     @property
     def num_frames(self):
         return len(self._src.pages)
+
+
+class ImageDescription:
+    def __init__(self, src):
+        try:
+            self.json = json.loads(src.description)
+        except json.decoder.JSONDecodeError:
+            self.json = {}
+            self._cmap = {}
+            return
+
+        self._cmap = {"red": 0, "green": 1, "blue": 2}
+
+        # update format if necessary
+        if "Alignment red channel" in self.json:
+            for color, j in self._cmap.items():
+                self.json[f"Channel {j} alignment"] = self.json.pop(f"Alignment {color} channel")
+                self.json[f"Channel {j} detection wavelength (nm)"] = "N/A"
+
+    def __bool__(self):
+        return bool(self.json)
+
+    @property
+    def alignment_roi(self):
+        return np.array(self.json["Alignment region of interest (x, y, width, height)"])
+
+    @property
+    def roi(self):
+        return np.array(self.json["Region of interest (x, y, width, height)"])
+
+    @property
+    def offsets(self):
+        return self.alignment_roi[:2] - self.roi[:2]
+
+    def _raw_alignment_matrix(self, color):
+        return np.array(self.json[f"Channel {self._cmap[color]} alignment"]).reshape((2, 3))
+
+    def alignment_matrix(self, color):
+        def correct_alignment_offset(alignment, x_offset, y_offset):
+            # translate the origin of the image so that it matches that of the original transform
+            translation = np.eye(3)
+            translation[0, -1] = -x_offset
+            translation[1, -1] = -y_offset
+            # apply the original transform to the translated image.
+            # it only needs to be resized from a 2x3 to a 3x3 matrix
+            original = np.vstack((alignment, [0, 0, 1]))
+            # translate the image back to the original origin.
+            # takes into account both the offset and the scaling performed by the first step
+            back_translation = np.eye(3)
+            back_translation[0, -1] = original[0, 0] * x_offset
+            back_translation[1, -1] = original[1, 1] * y_offset
+            # concatenate the transforms by multiplying their matrices and ignore the unused 3rd row
+            return np.dot(back_translation, np.dot(original, translation))[:2, :]
+
+        align_mat = self._raw_alignment_matrix(color)
+        x_offset, y_offset = self.offsets
+        if x_offset == 0 and y_offset == 0:
+            return align_mat
+        else:
+            return correct_alignment_offset(align_mat, x_offset, y_offset)
+
+    @property
+    def for_export(self):
+        out = copy(self.json)
+        if self:
+            for j in range(3):
+                out[f"Applied channel {j} alignment"] = out.pop(f"Channel {j} alignment")
+        return json.dumps(out, indent=4)